+++ tests/tests_unit/test_api/test_time_series.py
@@ -32,7 +32,7 @@ def mock_ts_response(rsps):
         ]
     }
     url_pattern = re.compile(
-        re.escape(TS_API._get_base_url_with_base_path()) + "/timeseries(?:/byids|/update|/delete|/list|/search|$|\?.+)"
+        re.escape(TS_API._get_base_url_with_base_path()) + r"/timeseries(?:/byids|/update|/delete|/list|/search|$|\?.+)"
     )
     rsps.assert_all_requests_are_fired = False
 
@@ -116,12 +116,17 @@ class TestTimeSeries:
 
     def test_delete_single(self, mock_ts_response):
         res = TS_API.delete(id=1)
-        assert {"items": [{"id": 1}]} == jsgz_load(mock_ts_response.calls[0].request.body)
+        assert {"items": [{"id": 1}], "ignoreUnknownIds": False} == jsgz_load(mock_ts_response.calls[0].request.body)
+        assert res is None
+
+    def test_delete_single_ignore_unknown(self, mock_ts_response):
+        res = TS_API.delete(id=1, ignore_unknown_ids=True)
+        assert {"items": [{"id": 1}], "ignoreUnknownIds": True} == jsgz_load(mock_ts_response.calls[0].request.body)
         assert res is None
 
     def test_delete_multiple(self, mock_ts_response):
         res = TS_API.delete(id=[1])
-        assert {"items": [{"id": 1}]} == jsgz_load(mock_ts_response.calls[0].request.body)
+        assert {"items": [{"id": 1}], "ignoreUnknownIds": False} == jsgz_load(mock_ts_response.calls[0].request.body)
         assert res is None
 
     def test_update_with_resource_class(self, mock_ts_response):